### 2019-09-24 #1758 @arfon
:wave: @statsmaths thanks for your submission to JOSS.

Could you suggest a few potential reviewers for your submission? This can be people you know of in the community, elsewhere on GitHub, or from [this list of potential reviewers](https://bit.ly/joss-reviewers).
