### 2019-09-30 #1758 @statsmaths
Of course. Here are some names:

- @duhaime **Douglas Duhaime**
- @adavidzh **André David**
- @SeguinBe **Benoit Seguin**
- @jgonggrijp **Julian Gonggrijp**
- @ZoeLeBlanc **Zoe LeBlanc**
