### 2019-10-01 #1758 @jgonggrijp
@arfon I would be willing, but I could only start November 11th at the earliest.

@statsmaths Out of curiosity, how did you find/know about me?
