### 2019-10-01 #1758 @arfon
> @arfon I would be willing, but I could only start November 11th at the earliest.

Thanks @jgonggrijp. Is that OK with you timing-wise?
