### 2019-10-01 #1758 @statsmaths
Waiting until 11 November would be fine. Maybe we could just wait a day to see if @duhaime might be willing (he reacted to the initial listing but is on New York time and likely has not seen the follow-up request yet).

@jgonggrijp Thanks! I found you through the UU DH Lab website. I work with several DH colleagues at Utrecht (e.g., Jasmijn van Gorp), know of the university's strength in visual DH, and was trying to find someone technical to review the work. When I saw that you have some repositories in Python and JavaScript, seemed like a good fit. Hopefully you don't mind!
