### 2019-10-01 #1758 @duhaime
@statsmaths @afron Thanks for the heads up on this! I wish I could help out, as the DVT will be useful to many researchers in @YaleDHLab, but my cup overfloweth at the moment. I'm glad @jgonggrijp can tackle this one!
