### 2019-10-06 #1758 @SeguinBe
Hello,

Sorry I am currently in holidays at the moment, so just seeing this
thread now, I will have a closer look at it around the 14th to see if
I could squeeze it in my agenda.

Le mer. 2 oct. 2019 à 14:14, Arfon Smith <notifications@github.com> a écrit :
>
> ⚡️ thanks for agreeing to review @jgonggrijp. I won't actually start the review issue here until we've found a second reviewer.
>
> @statsmaths - I'd like to find another reviewer for this submission. If you can think of any other names I'd be grateful - I'll also do some digging.
>
