### 2019-10-07 #1758 @arfon
:wave: @elektrobohemian - you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
