### 2019-10-10 #1758 @arfon
> @arfon I would be. starting around end of October/November would be great because I have to finish some Master theses before.

OK thanks. I think that timing works. I'll go ahead and set up the review now and set a reminder for the end of October.
