### 2019-10-10 #1800 @whedon
[REVIEW]: Distant Viewing Toolkit: A Python Package for the Analysis of Visual Culture
**Submitting author:** @statsmaths (<a href="http://orcid.org/0000-0003-0576-0669">Taylor Arnold</a>)
**Repository:** <a href="https://github.com/distant-viewing/dvt" target ="_blank">https://github.com/distant-viewing/dvt</a>
**Version:** v0.3.0
