### 2019-10-12 8b026d1
fix tensorflow and viz bugs
The new release of tensorflow (api 2) is now triggered
by the setup file; this has been changed back to 1.13
explicitly. A new version of the visualization engine has
also been added that deals with short video files that have
fewer than 5 shots.
