### 2019-11-12 #1800 @jgonggrijp
@arfon In my case, that was to be expected. As I mentioned [in the pre-review](https://github.com/openjournals/joss-reviews/issues/1758#issuecomment-536972108), I wouldn't be able to start before yesterday in any case.

I'll try to do the review within the next two weeks.
