### 2019-12-06 distant-viewing/dvt#12 @jgonggrijp
Question regarding JOSS paper authorship
@statsmaths @nolauren @Nanne

I'm currently reviewing the JOSS submission of `dvt` (https://github.com/openjournals/joss-reviews/issues/1800). One of the review criteria (per [JOSS documentation](https://joss.readthedocs.io/en/latest/review_criteria.html#authorship)) is that "the full list of paper authors seems appropriate and complete".

While I trust that the author list is correct, I am currently unable to objectively confirm this from the information available to me on GitHub:

- From the [paper](https://github.com/distant-viewing/dvt/blob/master/paper/paper.md), assuming I am looking at the right version, I find that the author list includes @nolauren and excludes @Nanne.
- From the [contributors graph](https://github.com/distant-viewing/dvt/graphs/contributors), the contributions by @Nanne *appear* to be more substantial than those by @nolauren. Both *appear* to be very small compared to the contributions by @statsmaths. Of course, I understand that this is only a measure of committed lines of code or text and that it completely disregards any contributions that may have been made outside of Git, including potentially important conceptual contributions.

Could you please clarify why @nolauren is included in the author list and why @Nanne is not? Please let me emphasize that I trust that there is a good reason; I'm just asking because this reason is currently invisible to me and I need it in order to confirm one of the review criteria. Thanks in advance.
