### 2019-12-06 distant-viewing/dvt#13 @jgonggrijp
Documented vs. actual license mismatch
This issue is part of my review in https://github.com/openjournals/joss-reviews/issues/1800.

In various forms of documentation, I find the consistent claim that `dvt` is licensed under GPL v2. I found this in the following places, but there may be more.

- https://github.com/distant-viewing/dvt/blob/master/paper/paper.md#process-and-development
- https://github.com/distant-viewing/dvt/blob/master/README.md (shield at the top)
- https://github.com/distant-viewing/dvt/blob/master/setup.py

The license that is actually included in the source code is the GPL v3: https://github.com/distant-viewing/dvt/blob/master/LICENSE.

Regarding the shield in the README, this could update automatically with the information available on PyPI if you were to use the following image URL (documentation and options [here](https://shields.io/category/license)):

Note that this reflects the information in the setup.py rather than the information in the actual license.
