### 2019-12-06 c833dc8
Make license choice of GPL-2 consistant
The LICENSE file included in the repository now reflect the GPL-2
license which was otherwise referenced elsewhere. This addresses
the mismatch from #13.
