### 2019-12-06 distant-viewing/dvt#13 @statsmaths
That is a good catch. It was intended to use GPL-2, but somehow included the wrong
LICENCE file. Not being a file that is opened very often, had not noticed this. Should
be fixed, along with the better PyPI link, in c833dc8.
