### 2019-12-06 distant-viewing/dvt#12 @statsmaths
Good question. We very much appreciate that @Nanne found the software useful and was willing to contribute two additional annotators to the toolkit. We hope that in the future other's will also want to contribute in this way as new methods and applications arise for the use of computer vision methods in the analysis of visual culture. The JOSS paper, however, is about the overall architecture and APIs exposed by the toolkit, which was solely the design and work of @nolauren and myself. Also, as you also mentioned, the contributes are a very limited proportion of the overall code-base.

The reason for including @nolauren as an author is two-fold. First, this is very much joint work that was initiated and produced collaboratively. We co-wrote and co-direct the lab and grant that supported the work, have co-run workshops focused on the design of the toolkit, and have each presented the toolkit and results using the toolkit at various conferences. Secondly, while the commit graph attributes nearly all of the code only to myself, many of the most important parts of the source code and documentation were actually written jointly in a pair-programming style while colocated at the University of Richmond. As the one with more Python experience, I was usually driving the computer and that's why most of the commits are tagged to me.

Hopefully that clarifies the disconnect between the commit graph and the authors. Please follow-up if more details are needed.
