### 2019-12-06 distant-viewing/dvt#12 @jgonggrijp
It does clarify and I'm satisfied. Thank you.
