### 2020-01-08 #1800 @statsmaths
Hi everyone. I know things get slowed down around the winter holidays, but wanted to check in now that things are starting to pick up again. I answered some issues that @jgonggrijp opened in the dvt repository, but had not seen anything since then.

If there are any issues people are running into, please let us know. If additional reviewers are needed I can try to recommend someone. Thanks again for everyone's help!
