### 2020-01-08 #1800 @arfon
Thanks for the update @statsmaths. I've also heard back over email from @elektrobohemian that they should be starting their review imminently.
