### 2020-01-08 #1800 @jgonggrijp
@statsmaths @arfon Just letting you know that this is still on my agenda. I hope to continue (and preferably finish) this Friday.
