### 2020-01-13 distant-viewing/dvt#15 @jgonggrijp
Comments on the documentation
This issue is part of my review in https://github.com/openjournals/joss-reviews/issues/1800. However, in this issue I'm only providing suggestions for minor improvements to the documentation. I consider all of these suggestions optional; the review criteria can also be met without adopting them.

- [x] The README and the Sphinx documentation both suggest that the user installs `dvt` in an Anaconda environment. I tried installing the package in a regular "bare" virtualenv instead and this worked as documented, too. So you could add a clarifying note that Anaconda is not strictly required. A regular virtualenv, or in fact any Python 3.7 environment with `pip` and a compiler toolchain to back it, should work.
- [x] On the page https://distant-viewing.github.io/dvt/tutorial/annotators.html, example code is shown. Following my programming habits, I copy-pasted the example code into a `.py` file, assuming that I would encounter a way to invoke the code from the command line later on. Instead, I found that the third code block already contained example output, so I was supposed to enter the code into an interactive Python console instead (this worked fine in my "bare" virtualenv, too). It would be helpful to mention this before introducing the example code. Another option is to include the `>>>` prompt in the example displays, following the pydoc/doctest convention. You already did this in the [DataExtraction API documentation](https://distant-viewing.github.io/dvt/core.html#dvt.core.DataExtraction). Doing both might be best.
- [x] The clarification from the previous point also applies to the [Low-level API section](https://github.com/distant-viewing/dvt#low-level-api-getting-started-with-the-python-interface) of the README.
- [x] While following the API tutorial, I got different results out of the `diff`, `demo_anno` and `demo_agg` examples. In `diff` and `demo_agg` the differences were all less than 1 and I'm generally not so impressed by differences in floating point numbers, but I found the differences in the `demo_anno` output quite apparent because I don't expect different machines to disagree about pixel values in a video file. Has the example video somehow been edited after the tutorials were written? Published example output [here](https://distant-viewing.github.io/dvt/tutorial/custom.html), my output below.
```py
dextra.get_data()['demo_anno'].head()
```
```
   frame  red  green  blue
0      0    6      1     0
1      1    6      1     0
2      2    6      1     0
3      3    6      1     0
4      4    4      2     0
```

- [x] In the [command line documentation page](https://distant-viewing.github.io/dvt/cli.html#), the text mentions an example file name of "input.mp4", but the example command has "video-clip.mp4". To me it was immediately clear that these refer to the same file, but to avoid confusion entirely, I suggest removing the discrepancy.
- [x] On the same page, there is an example command `python3 -m http.server –directory dvt-output-data` which has an N-dash unicode character instead of two regular dashes before the `directory` option. This will fail if users try to copy-paste the command. I don't know Sphinx very well, but I suspect this can be fixed by convincing Sphinx that the following line is code:
https://github.com/distant-viewing/dvt/blob/c833dc8661e3c754ddff4906a0e894ef3bc922fb/dvt/cli.py#L19
- [x] In the `VisualInput` abstract base class, documented [here](https://distant-viewing.github.io/dvt/abstract.html#dvt.abstract.VisualInput) and code referenced below, the `next_batch` method strongly reminds me of the special `__next__` method that is used in Python's iterator convention (documented [here](https://docs.python.org/3/library/stdtypes.html#typeiter)). Adopting this convention would permit more declarative `for batch in my_video` loops with `my_video` an instance of `VisualInput`.
https://github.com/distant-viewing/dvt/blob/c833dc8661e3c754ddff4906a0e894ef3bc922fb/dvt/abstract.py#L14-L55
- [x] In the [`dvt.utils.sub_image` documentation](https://distant-viewing.github.io/dvt/utils.html#dvt.utils.sub_image), the notion of a three-dimensional image deserves a bit more explanation. What is the third dimension? Color channels?
- [x] The [`dvt.annotate.cielab.CIElabAnnotator.num_buckets` documentation](https://distant-viewing.github.io/dvt/annotate/cielab.html#dvt.annotate.cielab.CIElabAnnotator.num_buckets) is very confusing because the documented default bucket size of 16 does not meet the documented requirement of being divisible by 256. Is it the other way round (should the bucket size be a divisor of 256, i.e., one of the first seven powers of 2) or does the requirement perhaps apply only when the bucket size exceeds some particular threshold?
- [x] In the documentation of [`dvt.annotate.diff.DiffAnnotator.size`](https://distant-viewing.github.io/dvt/annotate/diff.html#dvt.annotate.diff.DiffAnnotator.size), I presume the intention is “(Size of one side of the square) (used for down sampling [sic] the image)”, but I first read it as “(Size of one side of (the square used for down sampling [sic] the image))”, in which case it is ambiguous whether this square represents the target image size after resampling or a resampling window size (e.g. for a convolution). Perhaps it would be clearer to write “Side length of the square that the image will be downsampled to for the pixel-by-pixel comparison”.
- [x] The annotators section of the API documentation has two subsections that both have “Face Annotations” as the title. The [second](https://distant-viewing.github.io/dvt/annotate/img.html#module-dvt.annotate.img) actually describes the `ImgAnnotator`.
- [x] The API documentation of [`dvt.annotate.obj.ObjectAnnotator`](https://distant-viewing.github.io/dvt/annotate/obj.html#dvt.annotate.obj.ObjectAnnotator) has multiple occurrences of the word “face” that should probably read “object”.
