### 2020-01-13 distant-viewing/dvt#16 @jgonggrijp
State of the field
This issue is part of my review in https://github.com/openjournals/joss-reviews/issues/1800.

One of the requirements on the paper is to describe how your package compares to other commonly-used packages. Currently, the second paragraph of the summary mentions the existence of numerous standalone algorithm implementations but “limited options” for building end-to-end pipelines. It may be valuable to name these limited options explicitly and discuss their relative advantages and disadvantages.
