### 2020-01-13 distant-viewing/dvt#17 @jgonggrijp
Please provide instructions on how to run the tests
This issue is part of my review in https://github.com/openjournals/joss-reviews/issues/1800.

One of the review requirements is that there is a documented way to verify that the software works correctly. While there is a test suite, I was not able to find any instructions on how to run it. From searching and trial and error, it appears that these are the steps needed in order to run all tests with coverage and without errors after cloning the repository, although I still got a lot of warnings:

```
pip install .[tests,optional]
cd tests
pytest --cov=dvt
```

If installing from PyPI instead of cloning, substitute `dvt` for `.`.

Please include these instructions (or similar) in the Readme. It is also worth mentioning in the Readme that running all tests takes quite long; on my machine, it took nearly 15 minutes.
