### 2020-01-13 #1800 @jgonggrijp
@statsmaths @arfon I completed my review. I just opened three new issues (referenced right above this post), one of which is acceptance-blocking in my opinion (https://github.com/distant-viewing/dvt/issues/17) and one of which I'm not sure whether it should be acceptance-blocking or not (https://github.com/distant-viewing/dvt/issues/16). The latter might or might not be due to “limited options” being an euphemism for “no options”.

The third issue (https://github.com/distant-viewing/dvt/issues/15) is definitely non-blocking as far as I'm concerned. The package can be accepted with this issue still open.
