### 2020-01-15 eb0e34c
Adding test suite instructions to README
Solves issue #17. The code to run the test suite was only provided
in the travis.yml file. This adds explicit instructions to the
README file to make it easy to run by users not familiar with
travis or with pytest.
