+++ README.md
@@ -274,6 +274,23 @@ describing the tool and its application to the study of visual culture:
 }
 ```
 
+## Testing
+
+The distant viewing toolkit includes a full test suite, based on the python
+modules pytest. To run the tests clone the repository and run the following
+from a terminal. These commands install pytest and other optional dependencies,
+run the pytest unit tests, and display the coverage statistics:
+
+```
+pip install -U pytest
+pip install .[tests,optional]
+cd tests; pytest --disable-warnings --cov=dvt --cov-report term-missing .
+```
+
+To properly test the entire toolkit, we need to load and run several
+configurations of large neural network] models. Therefore, testing can take
+upwards of 20 minutes to complete.
+
 ## Contributing
 
 Contributions, including bug fixes and new features, to the toolkit are
