### 2020-01-15 distant-viewing/dvt#17 @statsmaths
Thanks for pointing the this. The code to run the test suite was in the `.travis.yml`
file, but that would not be an obvious place to look. I have included in commit
eb0e34c instructions in the README file and pointed out that the testing can take
upwards of 20 minutes because of the need to test various deep learning modules.
Finally, I also included the switch I use to silence the internal TensorFlow warnings.
