### 2020-01-15 a457d45
Adding references to paper
Addressing issue #16. Clarifying the state of the field of
end-to-end tools for processing of visual data and includes
references to three popular tools (PixPlot, ColorSpaces, and
FaceNet). Explains how DVT differs from these by offering a
wider range of methods within the pipline.
