+++ paper/paper.bib
@@ -48,12 +48,30 @@
   year      = "2018"
 }
 
+@misc{duhaime2020pixplot,
+  title = {{PixPlot}},
+  author = {Duhaime, Douglas},
+  year = {2019},
+  howpublished = {\url{https://github.com/YaleDHLab/pix-plot}},
+  note = {Accessed: 2020-01-15}
+}
+
 @inproceedings{farneback2003two,
   title     = "Two-frame motion estimation based on polynomial expansion",
   author    = "Farneb{\"a}ck, Gunnar",
   booktitle = "Scandinavian conference on Image analysis",
   pages     = "363--370",
   year      = "2003",
+}
+
+@article{flueckiger2018building,
+  title={Building a Crowdsourcing Platform for the Analysis of Film Colors},
+  author={Flueckiger, Barbara and Halter, Gaudenz},
+  journal={Moving Image: The Journal of the Association of Moving Image Archivists},
+  volume={18},
+  number={1},
+  pages={80--83},
+  year={2018}
 }
 
 @inproceedings{jiang2017face,
@@ -103,6 +121,14 @@
   booktitle = "Proccedings of the 3rd Workshop on Software Visualization, Evolution and Maintenance (VEM)",
   pages     = "33--40",
   year      = "2015"
+}
+
+@inproceedings{schroff2015facenet,
+  title={Facenet: A unified embedding for face recognition and clustering},
+  author={Schroff, Florian and Kalenichenko, Dmitry and Philbin, James},
+  booktitle={Proceedings of the IEEE conference on computer vision and pattern recognition},
+  pages={815--823},
+  year={2015}
 }
 
 @inproceedings{szegedy2017inception,
