### 2020-01-15 distant-viewing/dvt#16 @statsmaths
Thanks for pointing out this place for adding better references. The sentence now reads:

> However, there are limited options for users looking to quickly build end-to-end pipelines that link together common visual annotations. Those tools that do exist focus on specific subtasks, such as as FaceNet’s recognition of faces (Schroff, Kalenichenko, & Philbin, 2015), visual embedding with PixPlot (Duhaime, 2019), and FilmColors’ analysis of color (Flueckiger & Halter, 2018).
