### 2020-01-15 da18544
Updating documentation
Updates to documentation according to #15.
