### 2020-01-15 distant-viewing/dvt#15 @statsmaths
Thanks for all of the helpful comments about the documentation. Just made most of the
changes you recommended verbatim in da18544. Some details, particularly where I deviated
from your suggestions:

- The video test file was slightly changed because some users had trouble running it on
Windows, and that's what caused the issues with the discrepancy about the pixel values.
I have updated these.
- I am really glad to hear that it installed without any issues from a fresh virtual environment,
but choose to keep our recommendation of using Anaconda python. The reason for this is
that our target audience are digital humanists who often do not have compiler tools installed
and do not always know how to handle the error messages.
- I couldn't figure out how to get Sphinx to stop changing the double `--`, but changed the
command to make it work without that.
-  I made sure to separate anywhere that Sphinx combined code and output. I didn't include
the `>>>` because it makes it harder for people to copy paste (but I always go back and forth
about whether that's the best approach).
- The idea for the `VisualInput` iterator is a good one and something we will look into for the
next version. Users shouldn't be directly iterating over this, however, so it shouldn't effect the
end-user experience.
