### 2020-01-15 #1800 @statsmaths
@jgonggrijp Thanks for all of the really helpful comments. I pushed three commits that should
resolve the three issues that you opened.
