### 2020-01-15 #1800 @jgonggrijp
@statsmaths Welcome. dvt is a very interesting and well-documented package.
@arfon I see no blockers after the above fixes by @statsmaths.
