### 2020-01-20 #1800 @elektrobohemian
I really like the software and the documented installation process as it is very mature. This is why I tried to install the software without relying on my background language. Hence, all opened issues are non-blocking as far as I am concerned.
I took a basic Ubuntu 18.04.3 LTS installation and tried to solve all issues by copy-and-pasting all occuring issues. Under Mac OS, I tried to run everything in a zsh environment.

Besides the issues discussed in the dvt project, I would suggest to extend the paper a little. As the paper als addresses layperson users, you would like to be a little more informative regarding the local web server section. I guess that not all DH researches might be familiar with the Java script-related security policies of all browsers.

I strongly recommend the dvt package for publication. My minor issues could be resolved after publication.
