### 2020-01-20 #1800 @arfon
@statsmaths - looks like this is getting pretty close to being accepted here 😄. Can you confirm that you've made all of the changes you planned to based on the feedback from @jgonggrijp & @elektrobohemian
