### 2020-01-20 116ef22
improving documentation
Addressing issue #18, this makes it clear that version 3.7
of Python is required. It also modifies the testing code so
that it runs under bash and zsh (by escaping the brackets in
the pip command). It also further clarifies the role of
Anaconda Python: it is what is used primarily to test and
develope and we recommend it to users without much Python
experience. We also now mention that a virtual environemnt
under 3.7 should work as well.
