### 2020-01-20 b849dda
fixing documentation consistency
Fixes documentation issues in #20. Now mentions to download all
three files for the low-level tutorial, fixes the inconsistencies
on the README.md output (the video file was corrupted and fixing
it changes some of the output slightly), and fixes the link to the
Visual Style article
