### 2020-01-20 #1800 @statsmaths
Thanks @elektrobohemian for all of the helpful feedback. I just pushed responses that
resolve three of the four issues you opened. The only one still open is the issue with the
web application on macOS, as I cannot replicate it on my mac (it currently runs on Firefox,
Safari and Chrome for me). Not sure if that's blocking for the publication or not.

@arfon, other than a potential fix to the web server issue all of the changes have been
made at this point.
