### 2020-01-20 #1800 @elektrobohemian
regarding the web server: no, it's not as @jgonggrijp couldn't replicate it neither. at least in my opinion.

regarding the Ubuntu/virtualenv issue I wouldn't consider this blocking either as it would only happen to layperson users which might also not rely on a Ubuntu LTS. a layperson would prefer Anaconda IMHO. As said above, I was only checking these scenarios as everything else worked so fine.
