### 2020-01-20 b396f7b
update to 0.3.2 for release
