### 2020-01-20 #1800 @statsmaths
@arfon Great. Here is the DOI of the updated package. Thanks again to everyone!

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3614034.svg)](https://doi.org/10.5281/zenodo.3614034)
