### 2020-01-20 #1800 @arfon
@statsmaths - could you check if any of the DOIs that Whedon has found are correct? If they are, please go ahead and add them to your bibtex file (without the `https://doi.org/` preamble).
