### 2020-01-20 #1800 @arfon
@statsmaths - I think this PR is correct too: https://github.com/distant-viewing/dvt/pull/22
