### 2020-01-20 f0fe8ba
fixing bibliographic DOI entries
