+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @inproceedings{abadi2016tensorflow,
   title     = "Tensorflow: A system for large-scale machine learning",
   author    = "Abadi, Mart{\'\i}n and Barham, Paul and Chen, Jianmin and Chen, Zhifeng and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and Ghemawat, Sanjay and Irving, Geoffrey and Isard, Michael",
-  booktitle = "12th $\{$USENIX$\}$ Symposium on Operating Systems Design and Implementation ($\{$OSDI$\}$ 16)",
+  booktitle = "12th {USENIX} Symposium on Operating Systems Design and Implementation (OSDI} 16)",
   pages     = "265--283",
   year      = "2016"
 }
@@ -16,7 +16,7 @@
 }
 
 @article{arnold2019visual,
-  title   = "Visual Style in Two Network Era Sitcoms",
+  title   = "Visual Style in Two {N}etwork {E}ra Sitcoms",
   author  = "Arnold, Taylor B and Tilton, Lauren and Berke, Annie",
   journal = "Cultural Analytics",
   year    = "2019",
@@ -29,6 +29,7 @@
   author     = "Birnbaum, David J and Langmead, Alison",
   booktitle  = "New directions for computing education",
   pages      = "63--85",
+  doi        = "10.1007/978-3-319-54226-3_5",
   year       = "2017",
   publisher  = "Springer"
 }
@@ -41,10 +42,11 @@
 }
 
 @inproceedings{cao2018vggface2,
-  title     = "Vggface2: A dataset for recognising faces across pose and age",
+  title     = "{VggFace2}: A dataset for recognising faces across pose and age",
   author    = "Cao, Qiong and Shen, Li and Xie, Weidi and Parkhi, Omkar M and Zisserman, Andrew",
-  booktitle = "2018 13th IEEE International Conference on Automatic Face \& Gesture Recognition (FG 2018)",
+  booktitle = "2018 13th {IEEE} International Conference on Automatic Face \& Gesture Recognition",
   pages     = "67--74",
+  doi       = "10.1109/fg.2018.00020",
   year      = "2018"
 }
 
@@ -59,7 +61,8 @@
 @inproceedings{farneback2003two,
   title     = "Two-frame motion estimation based on polynomial expansion",
   author    = "Farneb{\"a}ck, Gunnar",
-  booktitle = "Scandinavian conference on Image analysis",
+  booktitle = "Scandinavian conference on image analysis",
+  doi       = "10.1007/3-540-45103-x_50",
   pages     = "363--370",
   year      = "2003",
 }
@@ -71,38 +74,42 @@
   volume={18},
   number={1},
   pages={80--83},
-  year={2018}
+  year={2018},
+  doi={10.5749/movingimage.18.1.0080}
 }
 
 @inproceedings{jiang2017face,
   title     = "Face detection with the faster R-CNN",
   author    = "Jiang, Huaizu and Learned-Miller, Erik",
-  booktitle = "2017 12th IEEE International Conference on Automatic Face \& Gesture Recognition (FG 2017)",
+  booktitle = "2017 12th {IEEE} International Conference on Automatic Face \& Gesture Recognition",
   pages     = "650--657",
+  doi       = "10.1109/fg.2017.82",
   year      = "2017"
 }
 
 @inproceedings{karasek2015color,
   title     = "Color image (dis) similarity assessment and grouping based on dominant colors",
   author    = "Karasek, Jan and Burget, Radim and Uher, Vaclav and Masek, Jan and Dutta, Malay Kishore",
-  booktitle = "2015 38th International Conference on Telecommunications and Signal Processing (TSP)",
[…]
