### 2020-01-20 #1800 @statsmaths
Okay, merged PR and put all the DOIs into the bibles file.
