### 2020-01-20 #1800 @arfon
@statsmaths - think this unmatched `}` is causing things to blow up: https://github.com/distant-viewing/dvt/blob/master/paper/paper.bib#L4
