+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @inproceedings{abadi2016tensorflow,
   title     = "Tensorflow: A system for large-scale machine learning",
   author    = "Abadi, Mart{\'\i}n and Barham, Paul and Chen, Jianmin and Chen, Zhifeng and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and Ghemawat, Sanjay and Irving, Geoffrey and Isard, Michael",
-  booktitle = "12th {USENIX} Symposium on Operating Systems Design and Implementation (OSDI} 16)",
+  booktitle = "12th {USENIX} Symposium on Operating Systems Design and Implementation ({OSDI} 16)",
   pages     = "265--283",
   year      = "2016"
 }
