### 2020-01-20 #1800 @statsmaths
Yes, your right (for some reason my pandoc compiled anyway). Fixed in 5f880da.
